can bus advanced as separate module
