+++ can_bus/advanced/can/bus/watchdog.go
@@ -1,9 +1,9 @@
 package bus
 
 import (
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/common"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/controller"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/physical"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/common"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/controller"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/physical"
 	"github.com/hovsep/fmesh/component"
 	"github.com/hovsep/fmesh/signal"
 )
