+++ can_bus/advanced/can/codec/bit_buffer.go
