+++ can_bus/advanced/can/common/common.go
