+++ can_bus/advanced/can/common/noop_logger.go
