+++ can_bus/advanced/can/controller/controller.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/codec"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/common"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/codec"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/common"
 	"github.com/hovsep/fmesh/signal"
 
 	"github.com/hovsep/fmesh/component"
