+++ can_bus/advanced/can/controller/tx_queue.go
@@ -1,6 +1,6 @@
 package controller
 
-import "github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/codec"
+import "github.com/hovsep/fmesh-example/can_bus/advanced/can/codec"
 
 type TxQueueItem struct {
 	Buf *codec.BitBuffer // Binary encoded frame, wih SOF, EOF, IFS and 1 extra bit
