+++ can_bus/advanced/can/node.go
@@ -1,10 +1,10 @@
 package can
 
 import (
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/bus"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/common"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/controller"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/microcontroller"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/bus"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/common"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/controller"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/microcontroller"
 	"github.com/hovsep/fmesh/component"
 )
