+++ can_bus/advanced/can/physical/voltage.go
@@ -1,6 +1,6 @@
 package physical
 
-import "github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/codec"
+import "github.com/hovsep/fmesh-example/can_bus/advanced/can/codec"
 
 type Voltage float64
