+++ can_bus/advanced/can/transceiver.go
@@ -3,10 +3,10 @@ package can
 import (
 	"errors"
 	"fmt"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/physical"
 
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/codec"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/common"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/codec"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/common"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/physical"
 	"github.com/hovsep/fmesh/component"
 	"github.com/hovsep/fmesh/signal"
 )
