+++ can_bus/advanced/diagnostics/laptop.go
@@ -1,8 +1,8 @@
 package diagnostics
 
 import (
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can"
 	"github.com/hovsep/fmesh-examples/can_bus/advanced/ecu/obd"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can"
 	"github.com/hovsep/fmesh/common"
 	"github.com/hovsep/fmesh/component"
 	"github.com/hovsep/fmesh/signal"
