--- can_bus/advanced/ecu/airbag.go
@@ -1,13 +1,0 @@
-package ecu
-
-import (
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can"
-)
-
-const (
-	ACUUnitName = "acu"
-)
-
-func NewACU() *can.Node {
-	return can.NewNode(ACUUnitName, nil, nil)
-}
