+++ can_bus/advanced/ecu/obd/obd.go
@@ -2,8 +2,9 @@ package obd
 
 import (
 	"errors"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/common"
+
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/common"
 	"github.com/hovsep/fmesh/component"
 	"github.com/hovsep/fmesh/port"
 )
