+++ can_bus/advanced/ecu/transmission/transmission.go
@@ -1,8 +1,8 @@
 package transmission
 
 import (
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/microcontroller"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/microcontroller"
 	"github.com/hovsep/fmesh/component"
 )
