+++ can_bus/advanced/go.mod
@@ -1,0 +1,15 @@
+module github.com/hovsep/fmesh-example/can_bus/advanced
+
+go 1.24.1
+
+require (
+	github.com/hovsep/fmesh v1.2.3-Artashat
+	github.com/hovsep/fmesh-examples v0.0.0-20250811142529-f26c975b3398
+	github.com/stretchr/testify v1.10.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
