+++ can_bus/advanced/main.go
@@ -4,15 +4,14 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/bus"
 	"github.com/hovsep/fmesh-examples/can_bus/advanced/diagnostics"
 	"github.com/hovsep/fmesh-examples/can_bus/advanced/ecu/engine"
 	"github.com/hovsep/fmesh-examples/can_bus/advanced/ecu/obd"
 	"github.com/hovsep/fmesh-examples/can_bus/advanced/ecu/transmission"
 
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/bus"
-
 	"github.com/hovsep/fmesh"
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can"
 )
 
 // This demo simulates a CAN bus system with a laptop connected via a USB–OBD interface.
