+++ can_bus/advanced/microcontroller/common.go
