+++ can_bus/advanced/microcontroller/mcu.go
@@ -1,7 +1,7 @@
 package microcontroller
 
 import (
-	"github.com/hovsep/fmesh-examples/can_bus/advanced/internal/can/common"
+	"github.com/hovsep/fmesh-example/can_bus/advanced/can/common"
 	"github.com/hovsep/fmesh/component"
 )
